+++ src/server/io.rs
@@ -1,0 +1,28 @@
+use super::stream::SwitchingTlsStream;
+use tokio::io::{AsyncRead, AsyncWrite};
+
+// Needed to swap out TcpStream for SwitchingTlsStream and vice versa.
+pub trait AsyncStream: AsyncRead + AsyncWrite + Send {}
+impl AsyncStream for tokio::net::TcpStream {}
+impl AsyncStream for SwitchingTlsStream {}
+
+pub trait Async2Stream: tokio02::io::AsyncRead + tokio02::io::AsyncWrite + Send + Unpin {}
+impl Async2Stream for tokio02::net::TcpStream {}
+impl Async2Stream for tokio02tls::TlsStream<tokio02::net::TcpStream> {}
+impl Async2Stream for tokio02tls::TlsStream<Box<dyn Async2Stream>> {}
+
+pub trait AsAsyncIo {
+    fn as_async_io(self) -> Box<dyn Async2Stream>;
+}
+
+impl AsAsyncIo for tokio02::net::TcpStream {
+    fn as_async_io(self) -> Box<dyn Async2Stream> {
+        Box::new(self)
+    }
+}
+
+impl AsAsyncIo for tokio02tls::TlsStream<Box<dyn Async2Stream>> {
+    fn as_async_io(self) -> Box<dyn Async2Stream> {
+        Box::new(self)
+    }
+}
